+++ package.json
@@ -53,6 +53,7 @@
     "axios": "^0.17.1",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^8.0.4",
+    "babel-plugin-add-module-exports": "^1.0.0",
     "bootstrap": "^4.1.3",
     "css-loader": "^1.0.0",
     "csvtojson": "^1.1.12",
