+++ src/layers/cubeicon-layer/index.ts
@@ -1,5 +1,7 @@
-import { Layer, LayerProps, AttributeManager } from 'deck.gl';
-import { GL, Model, CubeGeometry } from 'luma.gl';
+import { Layer } from '@deck.gl/core';
+import GL from 'luma.gl/constants';
+import { LayerProps, AttributeManager } from 'deck.gl';
+import { Model, CubeGeometry } from 'luma.gl';
 import vertex from './cubeicon-layer-vertex.glsl';
 import fragment from './cubeicon-layer-fragment.glsl';
 import { LightSettings } from '../../types';
