--- src/layers/enhanced-arc-layer/enhanced-arc-layer-vertex.glsl.ts
@@ -1,86 +1,0 @@
-// Inspired by arc-layer vertex shader in deck.gl
-
-export default `\
-#define SHADER_NAME enhanced-arc-layer-vertex-shader
-
-attribute vec3 positions;
-attribute vec4 instanceSourceColors;
-attribute vec4 instanceTargetColors;
-attribute vec4 instancePositions;
-attribute vec3 instancePickingColors;
-attribute float instanceStrokeWidths;
-
-
-uniform float numSegments;
-uniform vec2 viewportSize;
-uniform float opacity;
-uniform float renderPickingBuffer;
-
-varying vec4 vColor;
-
-float paraboloid(vec2 source, vec2 target, float ratio) {
-
-  vec2 x = mix(source, target, ratio);
-  vec2 center = mix(source, target, 0.5);
-
-  float dSourceCenter = distance(source, center);
-  float dXCenter = distance(x, center);
-  return (dSourceCenter + dXCenter) * (dSourceCenter - dXCenter);
-}
-
-// offset vector by strokeWidth pixels
-// offset_direction is -1 (left) or 1 (right)
-vec2 getExtrusionOffset(vec2 line_clipspace, float offset_direction, float strokeWidth) {
-  // normalized direction of the line
-  vec2 dir_screenspace = normalize(line_clipspace * viewportSize);
-  // rotate by 90 degrees
-  dir_screenspace = vec2(-dir_screenspace.y, dir_screenspace.x);
-
-  vec2 offset_screenspace = dir_screenspace * offset_direction * max(1.0, project_scale(strokeWidth)) / 2.0;
-  vec2 offset_clipspace = offset_screenspace / viewportSize * 2.0;
-
-  return offset_clipspace;
-}
-
-float getSegmentRatio(float index) {
-  return smoothstep(0.0, 1.0, index / (numSegments - 1.0));
-}
-
-vec3 getPos(vec2 source, vec2 target, float segmentRatio) {
-  float vertex_height = paraboloid(source, target, segmentRatio);
-
-  return vec3(
-    mix(source, target, segmentRatio),
-    sqrt(max(0.0, vertex_height))
-  );
-}
-
-void main(void) {
-  vec2 source = project_position(instancePositions.xy);
-  vec2 target = project_position(instancePositions.zw);
-
-  float segmentIndex = positions.x;
-  float segmentRatio = getSegmentRatio(segmentIndex);
-  // if it's the first point, use next - current as direction
-  // otherwise use current - prev
-  float indexDir = mix(-1.0, 1.0, step(segmentIndex, 0.0));
-  float nextSegmentRatio = getSegmentRatio(segmentIndex + indexDir);
-
-  vec3 currPos = getPos(source, target, segmentRatio);
-  vec3 nextPos = getPos(source, target, nextSegmentRatio);
-  vec4 curr = project_to_clipspace(vec4(currPos, 1.0));
-  vec4 next = project_to_clipspace(vec4(nextPos, 1.0));
-
-  // extrude
-  vec2 offset = getExtrusionOffset((next.xy - curr.xy) * indexDir, positions.y, instanceStrokeWidths);
-  gl_Position = curr + vec4(offset, 0.0, 0.0);
-
-  vec4 color = mix(instanceSourceColors, instanceTargetColors, segmentRatio) / 255.;
-
-  vColor = mix(
-    vec4(color.rgb, color.a * opacity),
-    vec4(instancePickingColors / 255., 1.),
-    renderPickingBuffer
-  );
-}
-`;
